requirements.txt
streamlit
openai
sqlalchemy
pandas
PyPDF2
nltk
uuid
python-dotenv